Tighten attendee types: extract to dedicated file, add PII blob validation
- Extract 9 type definitions from attendees.ts (813 lines) into attendee-types.ts
- Add isPiiBlob type guard replacing unsafe `as PiiBlob` assertion in parsePiiBlob
- parsePiiBlob now validates all required fields before trusting parsed JSON
- Re-export types from attendees.ts so existing consumers are unaffected
